# swaggerDocument.json
{
    "swagger": "2.0",
    "info": {
        "title": "배틀팡의 API 문서입니다.",
        "description": "with swagget document ",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/comments": {
            "get": {
                "x-swagger-router-controller": "comments",
                "tags": [
                    "댓글 관련 API"
                ],
                "description": "전체 배틀의 댓글 리스트를 취득합니다.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "comments to be returned",
                        "schema": {
                            "$ref": "#/definitions/comments"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            },
            "put": {
                "x-swagger-router-controller": "comments",
                "tags": [
                    "댓글 관련 API"
                ],
                "description": "put comments",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "comments to be returned",
                        "schema": {
                            "$ref": "#/definitions/comments"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
    }
}